Fix incorrect calculation of sequence frame count
The count was rounding down to the nearest second, so sub-1 second
sequences were considered to have 0 frames.

+++ model/src/main/kotlin/io/sparkled/model/util/SequenceUtils.kt
@@ -15,6 +15,6 @@ object SequenceUtils {
      * @return The number of frames available in the sequence.
      */
     fun getFrameCount(song: Song, sequence: Sequence): Int {
-        return song.getDurationMs()!! / MS_PER_SECOND * sequence.getFramesPerSecond()!!
+        return (song.getDurationMs()!! / MS_PER_SECOND.toFloat() * sequence.getFramesPerSecond()!!).toInt()
     }
 }
